requirements.txt
google-generativeai==0.8.5\npython-dotenv==1.1.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
